pyproject.toml
[project]
name = "ikigaiaikido"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Felipe Cabral", email = "felipe.benites@gmail.com"},
]
dependencies = [
    "jupyter>=1.1.1",
    "pandas>=2.2.3",
    "streamlit>=1.39.0",
    "streamlit-authenticator>=0.4.1",
    "PyYAML>=6.0.2",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
